backend/requirements.txt
# Core Django
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
drf-nested-routers==0.94.1

# Authentication
djangorestframework-simplejwt==5.3.1

# API Documentation
drf-spectacular==0.27.2
drf-nested-routers==0.94.1

# Database
psycopg2-binary==2.9.10

# Celery & Redis
celery==5.4.0
redis==5.2.1

# File Processing
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1

# Validation
pydantic==2.10.4
pydantic-settings==2.7.0

# Environment variables
python-dotenv==1.0.1

# Development
django-extensions==3.2.3
